Add directmessage handling

+++ src/main/java/org/kaschka/fersagers/discord/bot/utils/MessageUtils.java
@@ -1,5 +1,6 @@
 package org.kaschka.fersagers.discord.bot.utils;
 
+import net.dv8tion.jda.api.entities.ChannelType;
 import net.dv8tion.jda.api.entities.User;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 
@@ -13,7 +14,9 @@ public final class MessageUtils {
     }
 
     public static void logAndDeleteMessage(MessageReceivedEvent event) {
+        if(!event.isFromType(ChannelType.PRIVATE)) {
+            event.getMessage().delete().queue();
+        }
         logger.logChatMessage(event);
-        event.getMessage().delete().queue();
     }
 }
